Various changes. Fixing the data file generator. Remapped StationIDs from int > long but it causes problems. Started implementing Swagger related infrastructure

+++ tools/XmlGenerator/Models/chrFactions.cs
@@ -26,9 +26,9 @@ namespace EVEMon.XmlGenerator.Models
 
         public double? sizeFactor { get; set; }
 
-        public short? stationCount { get; set; }
+        public int? stationCount { get; set; }
 
-        public short? stationSystemCount { get; set; }
+        public int? stationSystemCount { get; set; }
 
         public int? militiaCorporationID { get; set; }
